feat(v3)!: typed asset diagnostics; HTML generation internalized; section 8 layout
