+++ lib/src/editor/controller.dart
