+++ lib/src/platform/web_asset_resolver.dart
