+++ test/core/monaco_controller_source_test.dart
@@ -4,9 +4,7 @@ import 'package:flutter_test/flutter_test.dart';
 
 void main() {
   test('commands await readiness uniformly on all platforms', () {
-    final source = File(
-      'lib/src/core/monaco_controller.dart',
-    ).readAsStringSync();
+    final source = File('lib/src/editor/controller.dart').readAsStringSync();
 
     // Protocol v3 removed the hand-rolled pre-ready queues and the web/native
     // divergence: every command gates on readiness through _ensureReady.
